change logic
